refactor: 관심사에 맞게 패키지 구조 변경
