+++ src/main/java/lotto/SumResult.java
@@ -1,3 +1,5 @@
+package lotto;
+
 public interface SumResult {
 
 	int sumRevenue();
