+++ src/test/java/lotto/LottoManagerTest.java
@@ -1,3 +1,5 @@
+package lotto;
+
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,7 +8,7 @@ class LottoManagerTest {
 
 	@Test
 	void calculateRevenueTest(){
-		assertThat(LottoManager.calculateRevenue( ()->10000, 15000))
+		assertThat(LottoManager.calculateRevenue(()->10000, 15000))
 			.isEqualTo(0.67); // 0.6666666에서 반올림
 	}
 }
