+++ src/test/java/lotto/LottoNumberTest.java
@@ -1,3 +1,5 @@
+package lotto;
+
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
