+++ src/test/java/lotto/LottoResultTest.java
@@ -1,11 +1,13 @@
+package lotto;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import type.LottoRewardType;
 
 import java.util.EnumMap;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoResultTest {
