+++ src/test/java/utils/DrawNumberTest.java
@@ -1,7 +1,10 @@
+package utils;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import utils.DrawNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
