+++ src/test/java/utils/NumberUtilsTest.java
@@ -6,13 +6,13 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import type.DecimalType;
 
 import java.util.stream.Stream;
 
-import static exception.LottoExceptionType.NEGATIVE_TEXT;
+import static type.LottoExceptionType.NEGATIVE_TEXT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.junit.jupiter.api.Assertions.*;
 
 class NumberUtilsTest {
