+++ src/test/java/utils/StringUtilsTest.java
@@ -1,18 +1,9 @@
 package utils;
 
-import exception.LottoException;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.MethodSource;
-
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static exception.LottoExceptionType.NEGATIVE_TEXT;
 
 class StringUtilsTest {
